scripts/requirements-image-retrieval.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia>=1.4.0
Pillow>=10.0.0
pymongo>=4.5.0
tqdm>=4.65.0
imagehash>=4.3.1 